+++ packages/bundler-plugin-core/src/errors/BadOIDCServiceError.ts
@@ -1,0 +1,5 @@
+export class BadOIDCServiceError extends Error {
+  constructor(msg: string, options?: ErrorOptions) {
+    super(msg, options);
+  }
+}