+++ packages/bundler-plugin-core/src/errors/InvalidSlugError.ts
@@ -1,5 +1,5 @@
 export class InvalidSlugError extends Error {
-  constructor(msg: string) {
-    super(msg);
+  constructor(msg: string, options?: ErrorOptions) {
+    super(msg, options);
   }
 }